Be explicit about type-conversion for QString::args()
Passing a quint32_le depended on a type conversion that might not be
obvious to the reader. That's going to stop working when
QString::arg() gets finnickier about which overloads get coerced to
integral types, to avoid problems with types that convert fo
floating-point but aren't recognised by the compiler *as*
floating-point, so were otherwise being cast via integral types and
losing precision.

Pick-to: 6.10
Task-number: QTBUG-138471

+++ src/qml/common/qv4compileddata.cpp
@@ -30,7 +30,7 @@ bool Unit::verifyHeader(QDateTime expectedSourceTimeStamp, QString *errorString)
 
     if (version != quint32(QV4_DATA_STRUCTURE_VERSION)) {
         *errorString = QString::fromUtf8("V4 data structure version mismatch. Found %1 expected %2")
-                               .arg(version, 0, 16).arg(QV4_DATA_STRUCTURE_VERSION, 0, 16);
+                               .arg(quint32(version), 0, 16).arg(QV4_DATA_STRUCTURE_VERSION, 0, 16);
         return false;
     }
